feat(security): implement Next.js 15 CSRF token generation and validation
- Add src/lib/csrf.ts with server-side CSRF protection functions
- generateCSRFToken(): generates 32-byte random tokens with automatic cookie setting
- validateCSRFToken(): validates tokens against cookies with timing-attack protection
- clearCSRFToken() and getCurrentCSRFToken() for token management
- Uses Next.js 15 cookies API for secure server-side-only operation
- Full TypeScript type safety with proper error handling
- Remove duplicated CSRF functions from security.ts to avoid client-side imports
